+++ usersec/rules.py
@@ -1,9 +1,15 @@
 import rules
 
+from adminsec.rules import is_hpcadmin
+
+
 # ------------------------------------------------------------------------------
 # Predicates
 # ------------------------------------------------------------------------------
-from adminsec.rules import is_hpcadmin
+
+
+# Without object
+# ------------------------------------------------------------------------------
 
 
 @rules.predicate
@@ -16,9 +22,46 @@ def _has_pending_group_request(user):
     return user.hpcgroupcreaterequest_requester.exists()
 
 
+# HpcUser object based
+# ------------------------------------------------------------------------------
+
+
+@rules.predicate
+def _is_hpcuser(user, hpcuser):
+    return hpcuser.user == user
+
+
+@rules.predicate
+def _is_pi_of_hpcuser(user, hpcuser):
+    owner = hpcuser.primary_group.owner
+
+    if owner:
+        return owner.user == user
+
+    return False
+
+
+@rules.predicate
+def _is_delegate_of_hpcuser(user, hpcuser):
+    delegate = hpcuser.primary_group.delegate
+
+    if delegate:
+        return delegate.user == user
+
+    return False
+
+
+# HpcGroupCreateRequest object based
+# ------------------------------------------------------------------------------
+
+
 @rules.predicate
 def _is_group_requester(user, hpcgroupcreaterequest):
     return hpcgroupcreaterequest.requester == user
+
+
+# HpcGroup object based
+# ------------------------------------------------------------------------------
 
 
 @rules.predicate
@@ -44,11 +87,16 @@ is_group_requester = ~is_hpcadmin & ~is_cluster_user & _is_group_requester
 is_group_member = ~is_hpcadmin & is_cluster_user & _is_group_member
 is_group_owner = ~is_hpcadmin & is_cluster_user & _is_group_owner
 is_group_delegate = ~is_hpcadmin & is_cluster_user & _is_group_delegate
+is_hpcuser = ~is_hpcadmin & is_cluster_user & _is_hpcuser
+is_pi_of_hpcuser = ~is_hpcadmin & is_cluster_user & _is_pi_of_hpcuser
+is_delegate_of_hpcuser = (
+    ~is_hpcadmin & is_cluster_user & _is_delegate_of_hpcuser
+)
 
 can_view_hpcgroup = is_group_owner | is_group_delegate | is_group_member
 can_view_hpcgroupcreaterequest = is_group_requester
 can_view_hpcgroupchangerequest = is_group_owner | is_group_delegate
-can_view_hpcuser = is_group_owner | is_group_delegate
+can_view_hpcuser = is_hpcuser | is_pi_of_hpcuser | is_delegate_of_hpcuser
 can_create_hpcgroupcreaterequest = is_orphan
 
 # ------------------------------------------------------------------------------